requirements.txt
bitarray==2.1.3
numpy==1.20.3
Pillow==8.2.0
scipy==1.7.0